pyproject.toml
[project]
name = "topodiffusionnet"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "blobfile>=3.2.0",
    "mpi4py>=4.1.1",
    "numpy>=2.4.2",
    "pillow>=12.1.1",
    "scipy>=1.17.0",
    "torch>=2.10.0",
]
